First draft of idioms (#2)
* Intro and motivation for idioms; begin listing basic idioms

* First draft in progress

* Update README.md

just a typo

* Update README.md

I changed the dimension `time` to `obs` that stands for observations. I think that's best since "time" is typically a physical variable with absolute values for Lagrangian data, rather than a dimension. And that's what we use for the GDP dataset for now in the NetCDF files.

Co-authored-by: Shane Elipot <selipot@miami.edu>